requirements.txt
uvicorn
fastapi
dotenv
requests
httpx 
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
